+++ database/migrations/2023_02_19_151649_create_profile_table.php
@@ -13,19 +13,21 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('profile', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('user_id');
-            $table->string('name');
-            $table->string('title')->nullable();
-            $table->string('email')->unique();
-            $table->string('phone')->nullable();
-            $table->string('avatar')->nullable();
-            $table->text('bio')->nullable();
+        if (!Schema::hasTable('profile')) {
+            Schema::create('profile', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('user_id');
+                $table->string('name');
+                $table->string('title')->nullable();
+                $table->string('email')->unique();
+                $table->string('phone')->nullable();
+                $table->string('avatar')->nullable();
+                $table->text('bio')->nullable();
 
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->timestamps();
-        });
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
